chore: stats library components (#717)
