+++ src/renderer/screens/Home/Events/Category.tsx
@@ -1,12 +1,15 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { AccordionItem, AccordionPanel } from '@/renderer/library/Accordion';
+import {
+  AccordionItem,
+  AccordionPanel,
+  AccordionCaretHeader,
+} from '@/renderer/library/Accordion';
 import { EventGroup } from './Wrappers';
 import { getEventChainId } from '@/utils/EventUtils';
 import { getCategory } from '@/config/chains';
 import { EventItem } from './EventItem';
-import { AccordionCaretHeader } from '@/renderer/library/Accordion/AccordionCaretHeaders';
 import type { EventCategoryProps } from './types';
 
 export const Category = ({
