+++ src/renderer/screens/Home/Manage/Permissions.tsx
@@ -7,8 +7,8 @@ import {
   Accordion,
   AccordionItem,
   AccordionPanel,
+  AccordionCaretSwitchHeader,
 } from '@/renderer/library/Accordion';
-import { AccordionCaretSwitchHeader } from '@app/library/Accordion/AccordionCaretHeaders';
 import { AccountsController } from '@/controller/renderer/AccountsController';
 import { ellipsisFn } from '@w3ux/utils';
 import { executeOneShot } from '@/renderer/callbacks/oneshots';
