+++ src/renderer/screens/Home/Manage/Wrappers.tsx
@@ -13,56 +13,6 @@ export const Wrapper = styled.div`
   max-height: 100%;
 `;
 
-export const HeadingWrapper = styled.div`
-  width: 100%;
-  padding: 0.5rem 0.75rem;
-  z-index: 3;
-  opacity: 0.75;
-  user-select: none;
-  cursor: pointer;
-
-  .flex {
-    display: flex;
-    column-gap: 0.5rem;
-    align-items: center;
-    padding: 0.25rem 0.5rem;
-    transition: background-color 0.15s ease-in-out;
-    border-bottom: 1px solid var(--border-secondary-color);
-
-    &:hover {
-      background-color: #141414;
-    }
-    > div {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      column-gap: 1rem;
-      padding: 0.5rem;
-    }
-
-    .left {
-      flex: 1;
-      display: flex;
-      column-gap: 0.75rem;
-      justify-content: flex-start;
-
-      .icon-wrapper {
-        min-width: 0.75rem;
-        opacity: 0.4;
-      }
-      h5 {
-        > span {
-          color: var(--text-color-primary);
-        }
-      }
-    }
-    .right {
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
-`;
-
 export const AccountsWrapper = styled.div`
   --item-height: 2.75rem;
   width: 100%;
