+++ src/renderer/screens/Home/Wrappers.ts
@@ -93,55 +93,3 @@ export const NoAccountsWrapper = styled.div`
     filter: brightness(140%);
   }
 `;
-
-export const HeadingWrapper = styled.div`
-  width: 100%;
-  margin-bottom: 1rem;
-  opacity: 0.75;
-  user-select: none;
-  cursor: pointer;
-
-  .flex {
-    padding: 0.25rem 0;
-    transition: background-color 0.15s ease-in-out;
-    border-bottom: 1px solid var(--border-secondary-color);
-    transition: background-color 0.15s ease-in-out;
-
-    &:hover {
-      background-color: #141414;
-    }
-    > div {
-      display: flex;
-      flex-direction: row;
-      align-items: baseline;
-      column-gap: 1rem;
-      padding: 0.5rem;
-    }
-    .left {
-      display: flex;
-      justify-content: flex-start;
-      flex: 1;
-
-      .icon-wrapper {
-        min-width: 0.75rem;
-        opacity: 0.4;
-      }
-    }
-    .right {
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
-
-  h5 {
-    display: flex;
-    align-items: flex-end;
-  }
-
-  .icon {
-    fill: var(--text-color-primary);
-    width: 0.95rem;
-    height: 0.95rem;
-    margin-right: 0.6rem;
-  }
-`;
